Revert changes to training edition and introduce dummy edition
